Code Clean up, plus new methods.
