+++ Methods/Parameters/Inventory/Inventory/Model/item.cs
@@ -1,21 +1,19 @@
-﻿
-
-namespace Inventory
+﻿namespace Inventory.Model
 {
     class Item
     {
-        public int Id { get;}
-        public string Name { get; set;}
+        public int Id { get; }
+        public string Name { get; set; }
         public int Quantity { get; set; }
         public int Price { get; set; }
-        
+
         public Item(string name, int quantity, int price)
         {
             Name = name;
-            Quantity = quantity; 
+            Quantity = quantity;
             Price = price;
         }
 
-       
+
     }
 }
